+++ main.go
@@ -11,7 +11,7 @@ import (
 )
 
 // version is overridden at release time via -ldflags "-X main.version=...".
-var version = "0.1.1"
+var version = "0.2.0"
 
 func main() {
 	if len(os.Args) < 2 {